Create a common utils folder
